Extracted getRelativeDir method to Codeception\Util\PathResolver (#3556)
